Add InventoryActivity
This activity displays the current list of items that the player owns
