+++ app/src/androidTest/java/ch/epfl/sweng/studyup/InventoryActivityTest.java
@@ -7,17 +7,17 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import ch.epfl.sweng.studyup.social.ChatActivity;
+import ch.epfl.sweng.studyup.items.InventoryActivity;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 
 @RunWith(AndroidJUnit4.class)
-public class ChatActivityTest {
+public class InventoryActivityTest {
     @Rule
-    public final ActivityTestRule<ChatActivity> rule =
-            new ActivityTestRule<>(ChatActivity.class);
+    public final ActivityTestRule<InventoryActivity> rule =
+            new ActivityTestRule<>(InventoryActivity.class);
 
     @Test
     public void testOptionNoException() {
