+++ app/src/main/AndroidManifest.xml
@@ -45,7 +45,7 @@
         <activity android:name=".player.QuestsActivityStudent" />
         <activity android:name=".teacher.QuestsActivityTeacher" />
         <activity android:name=".social.RankingsActivity" />
-        <activity android:name=".social.ChatActivity" />
+        <activity android:name=".items.InventoryActivity" />
         <activity android:name=".player.CustomActivity" />
 
         <!--
