+++ app/src/main/java/ch/epfl/sweng/studyup/items/DisplayItemActivity.java
@@ -1,0 +1,4 @@
+package ch.epfl.sweng.studyup.items;
+
+public class DisplayItemActivity {
+}
